Blockout Progress
Added the hub and part of the level blockout as well as made the level smallers. Cut off the top part into a smaller side and the bottom area into a thinner area. Created the red material for the chest. Still on work of implementing the beams and planks on the tunnels.